Add Dockerfile for a standalone FSO server
This is pretty bare bones but will launch FSO in standalone mode in a
container.

It expects a build parameter for the version and a volume pointing to a
FSO content root to be mounted at /fso.

FYI, the content root must contain a fs2_open.ini with at least a valid
VideocardFs2open value or SDL2 will cause a crash.